pyproject.toml
[project]
authors = [{name = "D'Angelo Salvatore", email = "you@example.com"}]
dependencies = [
  "ibm-vpc==0.29.1",
  "mcp[cli]>=1.10.1",
  "pandas==2.3.0",
  "pydantic>=2.11.7",
  "python-dotenv==1.1.1",
]
description = "Agentic tool to control IBM Cloud Virtual Servers (VSI)"
license = {text = "MIT"}
name = "ibm-cloud-vm-controller"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[dependency-groups]
dev = [
  "detect-secrets==1.5.0",
  "mypy>=1.16.1",
  "pre-commit==4.2.0",
  "pytest>=8.4.1",
  "flake8>=7.0.0",
  "black>=25.0.0",
  "isort>=5.13.0",
]

[tool.uv]
package = false

[tool.black]
line-length = 120

[tool.isort]
known_first_party = ["dtos", "services"]
line_length = 120
lines_between_sections = 0
profile = "black"

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
mypy_path = "src"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "ibm_cloud_sdk_core.authenticators"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "ibm_vpc"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "pandas"
